chore: move txtype-specific builders to network-primitives
