+++ crates/network/src/transaction/builder.rs
@@ -1,11 +1,11 @@
 use super::signer::NetworkWallet;
 use crate::Network;
-use alloy_consensus::BlobTransactionSidecar;
-use alloy_eips::eip7702::SignedAuthorization;
 use alloy_primitives::{Address, Bytes, ChainId, TxKind, U256};
 use alloy_rpc_types_eth::AccessList;
 use alloy_sol_types::SolCall;
 use futures_utils_wasm::impl_future;
+
+pub use alloy_network_primitives::{TransactionBuilder4844, TransactionBuilder7702};
 
 /// Result type for transaction builders
 pub type BuildResult<T, N> = Result<T, UnbuiltTransactionError<N>>;
@@ -342,49 +342,4 @@ pub trait TransactionBuilder<N: Network>: Default + Sized + Send + Sync + 'stati
         self,
         wallet: &W,
     ) -> impl_future!(<Output = Result<N::TxEnvelope, TransactionBuilderError<N>>>);
-}
-
-/// Transaction builder type supporting EIP-4844 transaction fields.
-pub trait TransactionBuilder4844: Default + Sized + Send + Sync + 'static {
-    /// Get the max fee per blob gas for the transaction.
-    fn max_fee_per_blob_gas(&self) -> Option<u128>;
-
-    /// Set the max fee per blob gas  for the transaction.
-    fn set_max_fee_per_blob_gas(&mut self, max_fee_per_blob_gas: u128);
-
-    /// Builder-pattern method for setting max fee per blob gas .
-    fn with_max_fee_per_blob_gas(mut self, max_fee_per_blob_gas: u128) -> Self {
-        self.set_max_fee_per_blob_gas(max_fee_per_blob_gas);
-        self
-    }
-
-    /// Gets the EIP-4844 blob sidecar of the transaction.
-    fn blob_sidecar(&self) -> Option<&BlobTransactionSidecar>;
-
-    /// Sets the EIP-4844 blob sidecar of the transaction.
-    ///
-    /// Note: This will also set the versioned blob hashes accordingly:
-    /// [BlobTransactionSidecar::versioned_hashes]
-    fn set_blob_sidecar(&mut self, sidecar: BlobTransactionSidecar);
-
-    /// Builder-pattern method for setting the EIP-4844 blob sidecar of the transaction.
-    fn with_blob_sidecar(mut self, sidecar: BlobTransactionSidecar) -> Self {
-        self.set_blob_sidecar(sidecar);
-        self
-    }
-}
-
-/// Transaction builder type supporting EIP-7702 transaction fields.
-pub trait TransactionBuilder7702: Default + Sized + Send + Sync + 'static {
-    /// Get the EIP-7702 authorization list for the transaction.
-    fn authorization_list(&self) -> Option<&Vec<SignedAuthorization>>;
-
-    /// Sets the EIP-7702 authorization list.
-    fn set_authorization_list(&mut self, authorization_list: Vec<SignedAuthorization>);
-
-    /// Builder-pattern method for setting the authorization list.
-    fn with_authorization_list(mut self, authorization_list: Vec<SignedAuthorization>) -> Self {
-        self.set_authorization_list(authorization_list);
-        self
-    }
 }
